#2 도메인 userId의 PK를 long으로 수정함에 따른 로직 변경

+++ src/main/java/University/Dormitory/Converter/WorkDateConverter.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 public class WorkDateConverter {
     @Builder
-    public static WorkDate SaveWorkDate(int userId, LocalDateTime startTime, LocalDateTime leaveTime) {
+    public static WorkDate SaveWorkDate(long userId, LocalDateTime startTime, LocalDateTime leaveTime) {
         return WorkDate.builder()
                 .scheduledLeaveTime(leaveTime)
                 .scheduledStartTime(startTime)
