+++ src/main/java/University/Dormitory/service/WorkService/WorkCommandService.java
@@ -14,7 +14,7 @@ public interface WorkCommandService {
      * @param wantStartTime 변경하려는 출근 시간
      *@param wantLeaveTime 변경하려는 퇴근 시간
      */
-    public String changeScheduleTimeByUserId(int userId,
+    public String changeScheduleTimeByUserId(long userId,
                                           LocalDateTime scheduledStartTime, LocalDateTime wantStartTime,
                                           LocalDateTime wantLeaveTime);
 
@@ -26,7 +26,7 @@ public interface WorkCommandService {
      * @param userId1Date 신청자의 줄 근무 시작 시간
      * return 값으로 종료시간과 시작시간 모두 주기. 그래야 명확하게 몇시부터 몇시 근무를 주는 지 알 수 있을 테니까.
      */
-    String giveMyWorkByUserIds(int userId1, int userId2, LocalDateTime userId1Date);
+    String giveMyWorkByUserIds(long userId1, long userId2, LocalDateTime userId1Date);
 
 
 //        근무 교체(나의 근무와 다른 사람의 근무 변경)
@@ -40,7 +40,7 @@ public interface WorkCommandService {
      * @param userId2StartTime 수락자의 근무 시작 시간
      * return 값으로 신청자의 시작 및 종료시간, 수락자의 신청 및 종료시간이 어떻게 바뀌는 지 알 수 있도록 시간 값 모두 주기(총 4개)
      */
-    String exchangeMyWorkByUserIds(int userId1, int userId2, LocalDateTime userId1StartTime, LocalDateTime userId2StartTime);
+    String exchangeMyWorkByUserIds(long userId1, long userId2, LocalDateTime userId1StartTime, LocalDateTime userId2StartTime);
 
 
     /**
@@ -50,7 +50,7 @@ public interface WorkCommandService {
      * @param reason 사유
      * @param date 사유 쓸 날짜. 이 날짜는 스케줄상의 날짜임.
      */
-    String writeReasonByUserId(int userId, String reason, LocalDateTime date);
+    String writeReasonByUserId(long userId, String reason, LocalDateTime date);
 
 
 
@@ -64,7 +64,7 @@ public interface WorkCommandService {
      * @param beforeActualTime 바꾸기 전의 시간
      * @param afterActualTime 바꿀 시간
      */
-    String changeActualTimeByUserId(int userId, LocalDateTime beforeActualTime, LocalDateTime afterActualTime);
+    String changeActualTimeByUserId(long userId, LocalDateTime beforeActualTime, LocalDateTime afterActualTime);
 
     /**
      * 처음 스케줄 짤 때 출/퇴근 넣기. 한번에 여러개일 경우 해당 함수 반복문 처리하면 될 것으로 예상.
@@ -74,7 +74,7 @@ public interface WorkCommandService {
      * @param leaveTime 근무 종료 시간
      */
 //    출/퇴근 근무시간 넣기 기능
-    String saveWorkByUserId(int userId, LocalDateTime startTime, LocalDateTime leaveTime);
+    String saveWorkByUserId(long userId, LocalDateTime startTime, LocalDateTime leaveTime);
 
 
     /**
@@ -84,7 +84,7 @@ public interface WorkCommandService {
      * @param date 우편근무일
      */
 //    우편 근무 날짜 넣기 기능
-    String postWorkByUserId(int userId, LocalDate date);
+    String postWorkByUserId(long userId, LocalDate date);
 
 
     /**
@@ -95,5 +95,5 @@ public interface WorkCommandService {
      * @param userId2 수락자의 학번
      * @param date2 수락자의 날짜
      */
-    String exchangePostWorkByUserId(int userId1, LocalDate date, int userId2, LocalDate date2);
+    String exchangePostWorkByUserId(long userId1, LocalDate date, long userId2, LocalDate date2);
 }
